pyproject.toml
[tool.poetry]
name = "uart-read"
version = "0.1.0"
description = ""
authors = ["raushanraja <vraushanraja@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyserial = "^3.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
